TEZ-4577: SortSpan could be created real small, resulting in eventual job failure

+++ tez-runtime-library/src/main/java/org/apache/tez/runtime/library/common/sort/impl/PipelinedSorter.java
@@ -947,7 +947,7 @@ public class PipelinedSorter extends ExternalSorter {
     public SortSpan(ByteBuffer source, int maxItems, int perItem, RawComparator comparator) {
       capacity = source.remaining();
       int metasize = METASIZE*maxItems;
-      int dataSize = maxItems * perItem;
+      long dataSize = (long) maxItems * (long) perItem;
       if(capacity < (metasize+dataSize)) {
         // try to allocate less meta space, because we have sample data
         metasize = METASIZE*(capacity/(perItem+METASIZE));
